src/db/migrations/20241223_01_DtThn-init.py
"""init."""

from yoyo import step

__depends__ = {}

steps = [
    step(
        """
            create table account
            (
                id   bigserial primary key,
                name text not null unique
            );
        """,
        """
            drop table account;
        """,
    ),
]
