requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiodns==3.2.0
    # via aiohttp
aiohttp[speedups]==3.9.5
    # via
    #   -r requirements.in
    #   aiohttp-apispec
    #   aiohttp-cors
    #   pytest-aiohttp
aiohttp-apispec==2.2.3
    # via -r requirements.in
aiohttp-cors==0.7.0
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
apispec==3.3.2
    # via aiohttp-apispec
async-timeout==4.0.3
    # via
    #   aiohttp
    #   asyncpg
asyncpg==0.30.0
    # via -r requirements.in
attrs==24.3.0
    # via aiohttp
brotli==1.1.0
    # via aiohttp
certifi==2024.12.14
    # via sentry-sdk
cffi==1.17.1
    # via pycares
coverage[toml]==7.6.9
    # via pytest-cov
exceptiongroup==1.2.2
    # via pytest
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
importlib-metadata==8.5.0
    # via yoyo-migrations
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via aiohttp-apispec
markupsafe==3.0.2
    # via jinja2
marshmallow==3.23.2
    # via
    #   -r requirements.in
    #   webargs
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
packaging==24.2
    # via
    #   marshmallow
    #   pytest
pluggy==1.5.0
    # via pytest
propcache==0.2.1
    # via yarl
psycopg2==2.9.10
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycares==4.5.0
    # via aiodns
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.20.1
    # via pydantic
pydantic-settings==2.7.0
    # via -r requirements.in
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-aiohttp
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
pytest-aiohttp==1.0.5
    # via -r requirements.in
pytest-asyncio==0.25.0
    # via pytest-aiohttp
pytest-cov==6.0.0
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via pydantic-settings
sentry-sdk==2.19.2
    # via -r requirements.in
sqlparse==0.5.3
    # via yoyo-migrations
structlog==24.4.0
    # via -r requirements.in
tabulate==0.9.0
    # via yoyo-migrations
tomli==2.2.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.12.2
    # via
    #   multidict
    #   pydantic
    #   pydantic-core
ujson==5.10.0
    # via -r requirements.in
urllib3==2.3.0
    # via sentry-sdk
uvloop==0.21.0
    # via -r requirements.in
webargs==5.5.3
    # via aiohttp-apispec
yarl==1.18.3
    # via aiohttp
yoyo-migrations==9.0.0
    # via -r requirements.in
zipp==3.21.0
    # via importlib-metadata
